+++ src/main/java/com/blackducksoftware/integration/hub/alert/web/controller/global/GlobalEmailConfigController.java
@@ -39,7 +39,7 @@ import com.blackducksoftware.integration.hub.alert.datasource.entity.global.Glob
 import com.blackducksoftware.integration.hub.alert.web.actions.global.GlobalEmailConfigActions;
 import com.blackducksoftware.integration.hub.alert.web.controller.CommonConfigController;
 import com.blackducksoftware.integration.hub.alert.web.controller.ConfigController;
-import com.blackducksoftware.integration.hub.alert.web.model.GlobalEmailConfigRestModel;
+import com.blackducksoftware.integration.hub.alert.web.model.global.GlobalEmailConfigRestModel;
 
 @RestController
 public class GlobalEmailConfigController extends ConfigController<GlobalEmailConfigRestModel> {
