+++ src/main/java/com/blackducksoftware/integration/hub/alert/web/model/global/GlobalEmailConfigRestModel.java
@@ -20,10 +20,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.blackducksoftware.integration.hub.alert.web.model;
+package com.blackducksoftware.integration.hub.alert.web.model.global;
 
 import org.apache.commons.lang3.builder.RecursiveToStringStyle;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
+
+import com.blackducksoftware.integration.hub.alert.web.model.ConfigRestModel;
 
 public class GlobalEmailConfigRestModel extends ConfigRestModel {
     private static final long serialVersionUID = 9172607945030111585L;
